Refactoring server codes
