+++ server/main.go
@@ -1,0 +1,13 @@
+package main
+
+import (
+	"github.com/gin-gonic/gin"
+	"github.com/illush123/hanamizuki/server/router"
+)
+
+func main() {
+	api := gin.Default()
+	router.Register(api)
+
+	api.Run(":8080")
+}
